put getField in a utils.js file, create a datatypes dir in src/ make it compile by gulp, move byte to numeric.js
